Extract TabbableContainer focus movement logic into extension function
